requirements.txt
# ============================================================================
# RTFM Discord Bot - Python Dependencies
# ============================================================================

# Discord Bot
discord.py==2.3.2

# Environment Management
python-dotenv==1.0.0

# Google Gemini API
google-generativeai==0.3.1

# LangChain for Vector DB
langchain==0.1.0
langchain-community==0.0.13

# Vector Database
chromadb==0.4.18

# Embeddings
sentence-transformers==2.2.2

# Required ML Dependencies
torch==2.1.1
numpy==1.26.2

# Kafka Integration
kafka-python==2.0.2
